camera: adjust camera hal to MM
LP camera hal has a reference to libion.so, although it is not using it.
libexynoscamera.so wants to use ion_alloc()/... of libion_exynos.so but
as libion.so (with the functions of the same name) is already
referenced, the functions of libion.so are erroneously used instead of
the libion_exynos.so ones.
This causes crashes whenever the camera is used (in ion_alloc).
Solution: replace libion.so with liblog.so which is already referenced by the lib (-> no-op)

Change-Id: I2a84d0d25013bcdc4ea9e3c09892ac3ee02bd266