libsec-ril: fix audio connection
In Android N the mediaserver process has been split into multiple
services (see https://nextiseasy.blogspot.de/2016/10/android-
nougat.html).
As a consequence the RIL audio connection is now initiated by
/system/bin/audioserver instead of /system/bin/mediaserver.

Samsung's libsec-ril checks if the process that connects to it via a
socket is a known process. The list of known processes only contains the
mediaserver but not audioserver process. Hence the audio connection is
denied ("RILD: can't connect with oem client").

To fix this, a new entry
<UID=0x411(=1041=AID_AUDIOSERVER),GID=-1(=ignore),Process="/system/bin/audioserver">
has been added to the allowedProcess list.

Change-Id: Ia8b3506758cc664a662c9f49cd93a69e310c69dd