proprietary: link libsec-ril.so against libprotobuf-cpp-fl26.so
Oreo comes with Google Protobuffer (libprotobuf-cpp-full.so) 3.0.
libsec-ril.so needs version 2.6. As other services need version 3.0 we
have to link 2.6 and 3.0 at the same time.
This is done by using a compatibility lib for version 2.6n(libprotobuf-
cpp-fl26.so). It requires that all namespaces are renamed to gxxgle.
For more info see binary_patcher.sh.

Change-Id: Ia99d3b35b4a1acd70b472521bcf98ec7968671f7